requirements.txt
streamlit==1.27.0
python-dotenv==1.0.1
google-api-core==2.12.0
langchain==0.1.236